move sbt.internal.util.Logger into sbt.util
