+++ internal/util-logging/src/main/scala/sbt/internal/util/FullLogger.scala
@@ -3,6 +3,8 @@
  */
 package sbt.internal.util
 
+import sbt.util._
+
 /** Promotes the simple Logger interface to the full AbstractLogger interface. */
 class FullLogger(delegate: Logger) extends BasicLogger {
   override val ansiCodesSupported: Boolean = delegate.ansiCodesSupported
