+++ internal/util-logging/src/main/scala/sbt/internal/util/GlobalLogging.scala
@@ -3,6 +3,7 @@
  */
 package sbt.internal.util
 
+import sbt.util._
 import java.io.{ File, PrintWriter }
 
 /**
