+++ internal/util-logging/src/main/scala/sbt/internal/util/MainLogging.scala
@@ -1,5 +1,6 @@
 package sbt.internal.util
 
+import sbt.util._
 import java.io.PrintWriter
 
 object MainLogging {
