+++ internal/util-logging/src/test/scala/TestLogger.scala
@@ -1,5 +1,7 @@
 package sbt.internal.util
 
+import sbt.util._
+
 object TestLogger {
   def apply[T](f: Logger => T): T =
     {
